+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Next
 
+### Bugfixes
+
 - Git plugin: package that changed `$Latest.PackageName` was not pushed when updated ([#66](https://github.com/majkinetor/au/issues/66)).
 
 ## 2017.3.29