+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # AU Project Changelog
 
-## Next
+## 2017.8.30
 
 - `Update-AUPackages` 
   - New options to handle update.ps1 errors: `IgnoreOn`, `RepeatOn`,`RepeatCount`,`RepeatSleep`. See [documentation](https://github.com/majkinetor/au#handling-update-errors). ([#76](https://github.com/majkinetor/au/issues/76))